backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pymongo[srv]
python-dotenv
pandas
transformers
torch
gunicorn
